Modify Tabular Data Dictionaries

Amends a tabular data dictionary, including adding/updating attributes and domains.

Usage

modify.table(
  x,
  how = c("add_attribute", "add_domain", "add_domainItem", "update_attribute",
    "update_domain", "update_domainItem"),
  codeName = "",
  allowNull = "",
  dataType = "",
  definition = "",
  units = "",
  unitsResolution = "",
  isCaseSensitive = "",
  missingValue = "",
  minValue = "",
  maxValue = "",
  fieldWidth = "",
  domainId = "",
  domainName = "",
  domainDescription = "",
  domainItem_value = "",
  domainItem_name = "",
  domainItem_definition = "",
  quiet = FALSE
)

Arguments

x

Data frame of a tabular data dictionary.

how

Character string matching one of the modification options: see ‘Details’.

codeName

Conditionally optional character string. Name of the attribute to add or modify.

allowNull

Optional logical (TRUE/FALSE). Whether null values are permissible for the attribute indicated with 'codeName'.

dataType

Optional character string. Datatype of the attribute indicated with 'codeName'.

definition

Optional character string. Definition of the attribute indicated with 'codeName'.

units

Optional character string. Unit-of-measure for the attribute indicated with 'codeName'.

unitsResolution

Optional numeric. Smallest increment of measurement for the attribute indicated with 'codeName'.

isCaseSensitive

Optional logical (TRUE/FALSE). Whether case-sensitive ASCII encoding is used for the attribute indicated with 'codeName'.

missingValue

Optional character string. Code representing missing values of the attribute indicated with 'codeName'.

minValue

Optional character string. Minimum value permissible for the attribute indicated with 'codeName'.

maxValue

Optional character string. Maximum value permissible for the attribute indicated with 'codeName'.

fieldWidth

Optional integer. Field width of the attribute indicated with 'codeName'.

domainId

Conditionally optional character string. Universally Unique Identifier (UUID) of the attribute or domain indicated with 'codeName' or domainName', respectively.

domainName

Conditionally optional character string. Name of the domain to add or modify.

domainDescription

Optional character string. Description of the domain indicated with 'domainName'.

domainItem_value

Optional character string. Value of the domain item (entry value) indicated with 'domainName'.

domainItem_name

Optional character string. Name of the domain item (entry value) indicated with 'domainItem_value'.

domainItem_definition

Optional character string. Definition of the domain item (entry value) indicated with 'domainItem_value'.

quiet

Optional logical. Default=FALSE. If TRUE, function runs with minimum input required and does not print revisions for approval.

Value

Returns a modified data frame of the data dictionary.

See also

Examples

TBD